finished/azure-container-apps/scale-container-aca/python/client/requirements.txt
azure-servicebus==7.12.0
azure-identity==1.15.0
